make test_sum faster on metal
This is a POC implementation of reduces using atomics. This works
for metal now.

Tested on test_sum which makes it 10x faster than it was. 80x to 8x
slower than pytorch (I ran it on not m1, so diff is so huge. Hopes this
works as well on apple silicon)