Create tinyrvc.py
bounty: voice changer model (so-vits-svc, rvc). make me sound like drake

...a beginning of sorts.

I plan to complete this bounty starting here at tailoring an rvc model and going from there to, hopefully, George doing a sick version of Started From the Bottom....

as a matter of fact, that is an appropriate theme for this code for me. 

please let me know if i am submitting this in a cringeworthy fashion and i'll fix things. 

i'm jason and hope to work with you all for the long haul.