Add a way to catch broken half handling in compiler-rt
compiler-rt 15 has a breaking ABI change affecting how half precision
floating point values are passed between functions
(https://github.com/llvm/llvm-project/issues/56854).

For reasons I don't fully understand, different versions of clang
exhibit the issue on different OSes. For example on Ubuntu 22.04
clang-15 is affected and clang-16 is OK, but on Fedora clang-16 is
affected. clang-16 is also the default on Fedora meaning that all the
half precision tests fail with random numbers because they're reading
garbage.

There's a one-liner to test for this issue in the LLVM project, so make
use of that to tell if the CLANG backend will produce correct results.