Work around compiler-rt half precision floating point ABI issues if we can
Taking a hint from https://github.com/llvm/llvm-project/issues/56854#issuecomment-1203573744
we can make use of `-march=native` flag to replace the library call
with a native instruction on Ivy Bridge and above. This doesn't fix the
bug in compiler-rt, but it means that we can support CLANG on more
hardware.

Note that the usual issues of `-march=native` shouldn't affect us since
we only compile these at runtime and we're building on the same machine
that we're running on. I've not measured if this improves the speed,
only that the tests now pass.