Quickstart: Upgrade section "Training" to new code
Upgrade the section "Training" in `docs/quickstart.md` to contain the
newest implementation of `sparse_categorical_crossentropy`, which has
not changed logically but is more concise than before.

Remove `indexing a tensor with another tensor` as an example of what
tinygrad does not support, since this is no longer true (see [1]) and
can also be a big disappointment for newcomers, since indexing a tensor
with another tensor is quite useful and common.

[1]: https://github.com/tinygrad/tinygrad/pull/1399