Add a new test to simulate typical scenario of foreign model being imported into the TG. We load the data as a stream of ints of various sizes, then bitcast them to match it original type and then often cast to some of our internal data types if needed.
That turns out to be a non-trivial task to get right, esp
dealing with historical backends with lots of hair and various
level of support for some of the types involved.

This test will excercise that conversion, the goal here is to
have most of the backends at least provide a consistent behaviour
when dealing with truncations and signedness.

Metal and OpenCL seems to be the platforms that are closer than
anyone else to the truth, OpenCL however also seems to vary
between different backends.