Fix softmax.argmax issue
argmax is finding max + finding argmax. The finding max part fused with softmax and introduced numerical error.
The resulting max can be different from any of the tensor element, and argmax returned size - 1 as a result.
One way to fix it is to include eps when comparing with float max.