use IGNORE_JIT_FIRST_BEAM to not BEAM in jit cnt=0
we wnat to have different BEAM value for resnet train and eval.
global JITBEAM cannot do this. added the flag to change beam behavior at cnt=0 (so it default behaves the same with or without TinyJit).

Also updated the context var BEAM in resnet to be outside of TinyJit. saves about 3 minutes compile time