Fix disk device not closing mmap
The disk device will now close the mmapped file when being closed.
This is needed on Windows where operations on the mmapped file will
result in "file in use" complaints.

This fixes some of the tests in `test_disk_tensor`, but modifications
to the tests were needed to make them work on Windows. Mainly freeing
the mmapped file before operating on it again in some cases.