Modify test_arange with new test sizes
Using 256 as our arange size caused GROUP operations on CPU platforms to
occasionally allow much better code generation than the 10x larger input
due to GROUPTOP operations having 64 and 256 as arguments in search.py.
This was not an issue as GROUP OptOps on GPU's modify the kernels
uniformly on either side allowing the flops to remain consistent. On
CPUs we implement GROUP's with multiple reduce axes which end up as for
loops. These are properly handled in using the loop collapse, though for
certain input sizes and OptOp configurations this meant a kernel was
emitted with vastly fewer flops, thus you wouldn't expect to keep a
linear relationship between input size and kernel flops in these cases.

So the simplest approach seems to be just tweaking the input values so
these best-case scenarios no longer happen and the test assumptions hold
true. I chose 768 as it equals 64*12 which gives us good combinations of
factors for potential OptOps for this test, while also removing the
chance that we get these best case optimization scenarios on the smaller
inputs.